requirements.txt
pymongo==4.3.3
motor==3.1.2
pydantic==1.10.7
fastapi==0.95.1
uvicorn==0.21.1
dnspython==2.3.0
python-dotenv==0.19.2
aiohttp==3.8.1
geopy==2.3.0